[ExtensionArray] Array에 extension 한것은 여기에 넣고 기존에 Extension은 ExtensionString으로 변경
